requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
python-dotenv
requests
aiosqlite